# Custom log filtering