--- a/src/ConfigCatClient/Evaluation/EvaluationDetails.cs
+++ b/src/ConfigCatClient/Evaluation/EvaluationDetails.cs
@@ -98,7 +98,7 @@ public abstract class EvaluationDetails
     /// </summary>
     public string? ErrorMessage
     {
-        get => this.errorMessage.ToString();
+        get => this.errorMessage.Value;
         set => this.errorMessage = value;
     }