--- a/src/ConfigCatClient/FormattableString.cs
+++ b/src/ConfigCatClient/FormattableString.cs
@@ -1,51 +1,18 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-// Source: https://github.com/dotnet/runtime/blob/v6.0.13/src/libraries/System.Private.CoreLib/src/System/FormattableString.cs
-
-#if NET45
+// Based on: https://github.com/dotnet/runtime/blob/v6.0.13/src/libraries/System.Private.CoreLib/src/System/FormattableString.cs
+// This is a modified version of the built-in type that is used only internally in the SDK.
 
 #pragma warning disable IDE0161 // Convert to file-scoped namespace
+#pragma warning disable CS0436 // Type conflicts with imported type
+
+global using ValueFormattableString = System.FormattableString;
 
 namespace System
 {
-    /// <summary>
-    /// A composite format string along with the arguments to be formatted. An instance of this
-    /// type may result from the use of the C# or VB language primitive "interpolated string".
-    /// </summary>
-    internal abstract class FormattableString : IFormattable
+    internal readonly struct FormattableString : IFormattable
     {
-        /// <summary>
-        /// The composite format string.
-        /// </summary>
-        public abstract string Format { get; }
-
-        /// <summary>
-        /// Returns an object array that contains zero or more objects to format. Clients should not
-        /// mutate the contents of the array.
-        /// </summary>
-        public abstract object?[] GetArguments();
-
-        /// <summary>
-        /// The number of arguments to be formatted.
-        /// </summary>
-        public abstract int ArgumentCount { get; }
-
-        /// <summary>
-        /// Returns one argument to be formatted from argument position <paramref name="index"/>.
-        /// </summary>
-        public abstract object? GetArgument(int index);
-
-        /// <summary>
-        /// Format to a string using the given culture.
-        /// </summary>
-        public abstract string ToString(IFormatProvider? formatProvider);
-
-        string IFormattable.ToString(string? ignored, IFormatProvider? formatProvider)
-        {
-            return ToString(formatProvider);
-        }
-
         /// <summary>
         /// Format the given object in the invariant culture. This static method may be
         /// imported in C# by
@@ -61,11 +28,6 @@ namespace System
         /// </summary>
         public static string Invariant(FormattableString formattable)
         {
-            if (formattable == null)
-            {
-                throw new ArgumentNullException(nameof(formattable));
-            }
-
             return formattable.ToString(Globalization.CultureInfo.InvariantCulture);
         }
 
@@ -84,19 +46,25 @@ namespace System
         /// </summary>
         public static string CurrentCulture(FormattableString formattable)
         {
-            if (formattable == null)
-            {
-                throw new ArgumentNullException(nameof(formattable));
-            }
-
             return formattable.ToString(Globalization.CultureInfo.CurrentCulture);
         }
 
-        public override string ToString()
+        private readonly string format;
+        private readonly object?[] arguments;
+
+        internal FormattableString(string format, object?[] arguments)
         {
-            return ToString(Globalization.CultureInfo.CurrentCulture);
+            this.format = format;
+            this.arguments = arguments;
         }
+
+        public string Format => this.format ?? string.Empty;
+        public object?[] GetArguments() { return this.arguments; }
+        public int ArgumentCount => this.arguments?.Length ?? 0;
+        public object? GetArgument(int index) { return this.arguments[index]; }
+
+        public string ToString(string? format, IFormatProvider? formatProvider) { return ToString(formatProvider); }
+        public string ToString(IFormatProvider? formatProvider) { return string.Format(formatProvider, this.format, this.arguments); }
+        public override string ToString() { return ToString(Globalization.CultureInfo.CurrentCulture); }
     }
 }
-
-#endif